+++ sdk/ai/azure-ai-agents/src/main/java/com/azure/ai/agents/implementation/AgentsClientImpl.java
@@ -81,6 +81,20 @@ public final class AgentsClientImpl {
     }
 
     /**
+     * The BetaAgentEndpointConversationsImpl object to access its operations.
+     */
+    private final BetaAgentEndpointConversationsImpl betaAgentEndpointConversations;
+
+    /**
+     * Gets the BetaAgentEndpointConversationsImpl object to access its operations.
+     * 
+     * @return the BetaAgentEndpointConversationsImpl object.
+     */
+    public BetaAgentEndpointConversationsImpl getBetaAgentEndpointConversations() {
+        return this.betaAgentEndpointConversations;
+    }
+
+    /**
      * The BetaMemoryStoresImpl object to access its operations.
      */
     private final BetaMemoryStoresImpl betaMemoryStores;
@@ -120,6 +134,20 @@ public final class AgentsClientImpl {
      */
     public AgentsImpl getAgents() {
         return this.agents;
+    }
+
+    /**
+     * The AgentTelephoniesImpl object to access its operations.
+     */
+    private final AgentTelephoniesImpl agentTelephonies;
+
+    /**
+     * Gets the AgentTelephoniesImpl object to access its operations.
+     * 
+     * @return the AgentTelephoniesImpl object.
+     */
+    public AgentTelephoniesImpl getAgentTelephonies() {
+        return this.agentTelephonies;
     }
 
     /**
@@ -184,9 +212,11 @@ public final class AgentsClientImpl {
         this.serializerAdapter = serializerAdapter;
         this.endpoint = endpoint;
         this.serviceVersion = serviceVersion;
+        this.betaAgentEndpointConversations = new BetaAgentEndpointConversationsImpl(this);
         this.betaMemoryStores = new BetaMemoryStoresImpl(this);
         this.betaAgents = new BetaAgentsImpl(this);
         this.agents = new AgentsImpl(this);
+        this.agentTelephonies = new AgentTelephoniesImpl(this);
         this.toolboxes = new ToolboxesImpl(this);
     }
 }